Implemented the Audio module (WIP)

+++ pyxel/core/include/pyxelcore/oscillator.h
@@ -17,21 +17,21 @@ class Oscillator {
 
  private:
   int32_t phase_;
-  int32_t (Oscillator::*tone_)(int32_t period, int32_t phase);
+  float (Oscillator::*tone_)(int32_t period, int32_t phase);
   int32_t period_;
   int32_t volume_;
 
-  int32_t (Oscillator::*next_tone_)(int32_t period, int32_t phase);
+  float (Oscillator::*next_tone_)(int32_t period, int32_t phase);
   int32_t next_period_;
   int32_t next_volume_;
 
   int32_t noise_seed_;
   int32_t noise_last_;
 
-  int32_t Triangle(int32_t period, int32_t phase);
-  int32_t Square(int32_t period, int32_t phase);
-  int32_t Pulse(int32_t period, int32_t phase);
-  int32_t Noise(int32_t period, int32_t phase);
+  float Triangle(int32_t period, int32_t phase);
+  float Square(int32_t period, int32_t phase);
+  float Pulse(int32_t period, int32_t phase);
+  float Noise(int32_t period, int32_t phase);
 };
 
 inline Oscillator::Oscillator() {
@@ -95,28 +95,28 @@ inline int32_t Oscillator::Output() {
   return output;
 }
 
-inline int32_t Oscillator::Triangle(int32_t period, int32_t phase) {
+inline float Oscillator::Triangle(int32_t period, int32_t phase) {
   float x = static_cast<float>(phase) / period + 0.25f;
   x -= static_cast<int32_t>(x);
 
   return Abs(x * 4.0f - 2.0f) - 1.0f;
 }
 
-inline int32_t Oscillator::Square(int32_t period, int32_t phase) {
+inline float Oscillator::Square(int32_t period, int32_t phase) {
   float x = static_cast<float>(phase) / period;
   x -= static_cast<int32_t>(x);
 
   return (x < 0.5f ? 1.0f : -1.0f) * 0.2f;
 }
 
-inline int32_t Oscillator::Pulse(int32_t period, int32_t phase) {
+inline float Oscillator::Pulse(int32_t period, int32_t phase) {
   float x = static_cast<float>(phase) / period;
   x -= static_cast<int32_t>(x);
 
   return (x < 0.25f ? 1.0f : -1.0f) * 0.2f;
 }
 
-inline int32_t Oscillator::Noise(int32_t period, int32_t phase) {
+inline float Oscillator::Noise(int32_t period, int32_t phase) {
   if (phase % (period / 4) == 0) {
     noise_seed_ >>= 1;
     noise_seed_ |= ((noise_seed_ ^ (noise_seed_ >> 1)) & 1) << 15;
